Verify that numpy types work with fastapi.

+++ api_params.py
@@ -1,8 +1,10 @@
+from numpy import float64
 from pydantic import BaseModel
 
 
 class Parameters(BaseModel):
-	sepal_length: float
-	sepal_width: float
-	petal_length: float
-	petal_width: float
+	sepal_length: float64
+	sepal_width: float64
+	petal_length: float64
+	petal_width: float64
+
